pyproject.toml
[project]
name = "qbfuck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-qbittorrent>=0.4.3",
]
